APTDecoder

APTDecoder

APTDecoder Julia package repo.

APTDecoder.decode — Method.

datatime, (channelA,channelB), data = decode(y,Fs)

Decode the APT image in a time series y defined at a frequency Fs (in Hz). datatime is the time in seconds counting from the beginning of the recording. channelA and channelB are arrays representing the data from the two different channels (A and B). data is raw data as an array including channel A and B, as well sync markers and telemetry.

Example

using FileIO
using APTDecoder

wavname = "gqrx_20190804_141523_137100000.wav"
y,Fs,nbits,opt = load(wavname)
datatime,(channelA,channelB),data = APTDecode.decode(y,Fs)

source

APTDecoder.georeference — Method.

plon,plat,data = georeference(pngname,satellite_name,channel)

Compute longitude and latitude of the NOAA APT satellite image in pngname using the orbit of the satellite with the name satellite_name (generally "NOAA 15", "NOAA 18", "NOAA 19"). The file name pngname should have the followng structure: string_date_time_frequency.png like gqrx_20190811_075102_137620000.png. Date and time of the file name are in UTC.

Example:

satellite_name = "NOAA 15"
pngname = "gqrx_20190811_075102_137620000.png";
APTDecoder.georeference(pngname,satellite_name)

source

APTDecoder.landseamask — Method.

lon,lat,data = APTDecoder.landseamask(;resolution='l',grid=5)

Load the land-sea-lake raster from basemap: 0 is ocean, 1 is land and 2 is lake. grid is the resolution in arc seconds and should be either 1.25, 2.5, 5 or 10. The parameter resolution should be either 'c','l','i','h' or 'f'.

The data is downloaded from: https://github.com/matplotlib/basemap/tree/master/lib/mpl_toolkits/basemap/data

The data is originally from GMT and distributed under the terms of the GPL licences.

using PyPlot
lon,lat,data = APTDecoder.landseamask(;resolution='c',grid=5)
pcolormesh(lon,lat,data'); colorbar()

source

APTDecoder.makeplots — Method.

makeplots(wavname,satellite_name)

Decodes the APT signal in wavname as recorded by gqrx using wide FM-mono demodulation. The file name wavname should have the followng structure: string_date_time_frequency.wav like gqrx_20190811_075102_137620000.wav. Date and time of the file name are in UTC (not local time). satellite_name is the name of the satellite (generally "NOAA 15", "NOAA 18", "NOAA 19").

Example:

import APTDecoder

wavname = "gqrx_20190825_182745_137620000.wav"
APTDecoder.makeplots(wavname,"NOAA 15")

source